requirements.txt
openenv-core>=0.2.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
websockets>=15.0.0
openai>=1.30.0
mistralai>=1.0.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
